refactor(elb): move elb resources to services path
