+++ huaweicloud/provider.go
@@ -606,15 +606,15 @@ func Provider() *schema.Provider {
 			"huaweicloud_drs_job":     drs.ResourceDrsJob(),
 			"huaweicloud_dws_cluster": dws.ResourceDwsCluster(),
 
-			"huaweicloud_elb_certificate":  ResourceCertificateV3(),
-			"huaweicloud_elb_l7policy":     ResourceL7PolicyV3(),
-			"huaweicloud_elb_l7rule":       ResourceL7RuleV3(),
+			"huaweicloud_elb_certificate":  elb.ResourceCertificateV3(),
+			"huaweicloud_elb_l7policy":     elb.ResourceL7PolicyV3(),
+			"huaweicloud_elb_l7rule":       elb.ResourceL7RuleV3(),
 			"huaweicloud_elb_listener":     elb.ResourceListenerV3(),
 			"huaweicloud_elb_loadbalancer": elb.ResourceLoadBalancerV3(),
-			"huaweicloud_elb_monitor":      ResourceMonitorV3(),
-			"huaweicloud_elb_ipgroup":      ResourceIpGroupV3(),
-			"huaweicloud_elb_pool":         ResourcePoolV3(),
-			"huaweicloud_elb_member":       ResourceMemberV3(),
+			"huaweicloud_elb_monitor":      elb.ResourceMonitorV3(),
+			"huaweicloud_elb_ipgroup":      elb.ResourceIpGroupV3(),
+			"huaweicloud_elb_pool":         elb.ResourcePoolV3(),
+			"huaweicloud_elb_member":       elb.ResourceMemberV3(),
 
 			"huaweicloud_enterprise_project": eps.ResourceEnterpriseProject(),
