+++ huaweicloud/services/acceptance/elb/resource_huaweicloud_elb_member_test.go
@@ -1,0 +1,234 @@
+package elb
+
+import (
+	"fmt"
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
+
+	"github.com/chnsz/golangsdk/openstack/elb/v3/pools"
+	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/config"
+	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/acceptance"
+	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/utils/fmtp"
+)
+
+func TestAccElbV3Member_basic(t *testing.T) {
+	var member_1 pools.Member
+	var member_2 pools.Member
+	rName := fmt.Sprintf("tf-acc-test-%s", acctest.RandString(5))
+
+	resource.ParallelTest(t, resource.TestCase{
+		PreCheck:          func() { acceptance.TestAccPreCheck(t) },
+		ProviderFactories: acceptance.TestAccProviderFactories,
+		CheckDestroy:      testAccCheckElbV3MemberDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccElbV3MemberConfig_basic(rName),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckElbV3MemberExists("huaweicloud_elb_member.member_1", &member_1),
+					testAccCheckElbV3MemberExists("huaweicloud_elb_member.member_2", &member_2),
+				),
+			},
+			{
+				Config: testAccElbV3MemberConfig_update(rName),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("huaweicloud_elb_member.member_1", "weight", "10"),
+					resource.TestCheckResourceAttr("huaweicloud_elb_member.member_2", "weight", "15"),
+				),
+			},
+		},
+	})
+}
+
+func testAccCheckElbV3MemberDestroy(s *terraform.State) error {
+	config := acceptance.TestAccProvider.Meta().(*config.Config)
+	elbClient, err := config.ElbV3Client(acceptance.HW_REGION_NAME)
+	if err != nil {
+		return fmtp.Errorf("Error creating HuaweiCloud elb client: %s", err)
+	}
+
+	for _, rs := range s.RootModule().Resources {
+		if rs.Type != "huaweicloud_elb_member" {
+			continue
+		}
+
+		poolId := rs.Primary.Attributes["pool_id"]
+		_, err := pools.GetMember(elbClient, poolId, rs.Primary.ID).Extract()
+		if err == nil {
+			return fmtp.Errorf("Member still exists: %s", rs.Primary.ID)
+		}
+	}
+
+	return nil
+}
+
+func testAccCheckElbV3MemberExists(n string, member *pools.Member) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		rs, ok := s.RootModule().Resources[n]
+		if !ok {
+			return fmtp.Errorf("Not found: %s", n)
+		}
+
+		if rs.Primary.ID == "" {
+			return fmtp.Errorf("No ID is set")
+		}
+
+		config := acceptance.TestAccProvider.Meta().(*config.Config)
+		elbClient, err := config.ElbV3Client(acceptance.HW_REGION_NAME)
+		if err != nil {
+			return fmtp.Errorf("Error creating HuaweiCloud elb client: %s", err)
+		}
+
+		poolId := rs.Primary.Attributes["pool_id"]
+		found, err := pools.GetMember(elbClient, poolId, rs.Primary.ID).Extract()
+		if err != nil {
+			return err
+		}
+
+		if found.ID != rs.Primary.ID {
+			return fmtp.Errorf("Member not found")
+		}
+
+		*member = *found
+
+		return nil
+	}
+}
+
+func testAccElbV3MemberConfig_basic(rName string) string {
+	return fmt.Sprintf(`
+data "huaweicloud_vpc_subnet" "test" {
+  name = "subnet-default"
+}
+
+data "huaweicloud_availability_zones" "test" {}
+
+resource "huaweicloud_elb_loadbalancer" "test" {
+  name            = "%s"
+  ipv4_subnet_id  = data.huaweicloud_vpc_subnet.test.subnet_id
+  ipv6_network_id = data.huaweicloud_vpc_subnet.test.id
+
+  availability_zone = [
+    data.huaweicloud_availability_zones.test.names[0]
+  ]
+}
+
+resource "huaweicloud_elb_listener" "test" {
+  name            = "%s"
+  description     = "test description"
+  protocol        = "HTTP"
+  protocol_port   = 8080
+  loadbalancer_id = huaweicloud_elb_loadbalancer.test.id
+
+  forward_eip = true
+
+  idle_timeout = 60
+  request_timeout = 60
+  response_timeout = 60
+}
+
+resource "huaweicloud_elb_pool" "test" {
+  name        = "%s"
+  protocol    = "HTTP"
+  lb_method   = "ROUND_ROBIN"
+  listener_id = huaweicloud_elb_listener.test.id
+}
+
+resource "huaweicloud_elb_member" "member_1" {
+  address       = "192.168.0.10"
+  protocol_port = 8080
+  pool_id       = huaweicloud_elb_pool.test.id
+  subnet_id     = data.huaweicloud_vpc_subnet.test.subnet_id
+
+  timeouts {
+    create = "5m"
+    update = "5m"
+    delete = "5m"
+  }
+}
+
+resource "huaweicloud_elb_member" "member_2" {
+  address       = "192.168.0.11"
+  protocol_port = 8080
+  pool_id       = huaweicloud_elb_pool.test.id
+  subnet_id     = data.huaweicloud_vpc_subnet.test.subnet_id
+
+  timeouts {
+    create = "5m"
+    update = "5m"
+    delete = "5m"
+  }
+}
+`, rName, rName, rName)
+}
+
+func testAccElbV3MemberConfig_update(rName string) string {
+	return fmt.Sprintf(`
+data "huaweicloud_vpc_subnet" "test" {
+  name = "subnet-default"
+}
+
+data "huaweicloud_availability_zones" "test" {}
+
+resource "huaweicloud_elb_loadbalancer" "test" {
+  name            = "%s"
+  ipv4_subnet_id  = data.huaweicloud_vpc_subnet.test.subnet_id
+  ipv6_network_id = data.huaweicloud_vpc_subnet.test.id
+
+  availability_zone = [
+    data.huaweicloud_availability_zones.test.names[0]
+  ]
+}
+
+resource "huaweicloud_elb_listener" "test" {
+  name            = "%s"
+  description     = "test description"
+  protocol        = "HTTP"
+  protocol_port   = 8080
+  loadbalancer_id = huaweicloud_elb_loadbalancer.test.id
+
+  forward_eip = true
+
+  idle_timeout = 60
+  request_timeout = 60
+  response_timeout = 60
+}
+
+resource "huaweicloud_elb_pool" "test" {
+  name        = "%s"
+  protocol    = "HTTP"
+  lb_method   = "ROUND_ROBIN"
+  listener_id = huaweicloud_elb_listener.test.id
+}
+
+resource "huaweicloud_elb_member" "member_1" {
+  address        = "192.168.0.10"
+  protocol_port  = 8080
+  weight         = 10
+  pool_id        = huaweicloud_elb_pool.test.id
+  subnet_id      = data.huaweicloud_vpc_subnet.test.subnet_id
+
+  timeouts {
+    create = "5m"
+    update = "5m"
+    delete = "5m"
+  }
+}
+
+resource "huaweicloud_elb_member" "member_2" {
+  address        = "192.168.0.11"
+  protocol_port  = 8080
+  weight         = 15
+  pool_id        = huaweicloud_elb_pool.test.id
+  subnet_id      = data.huaweicloud_vpc_subnet.test.subnet_id
+
+  timeouts {
+    create = "5m"
+    update = "5m"
+    delete = "5m"
+  }
+}
+`, rName, rName, rName)
+}
